Rename command classes
