+++ tests/Unit/ServeTest.php
@@ -7,21 +7,22 @@ use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\File;
 use Scabbard\Tests\TestCase;
 
-class WatchSiteTest extends TestCase
+class ServeTest extends TestCase
 {
-    public function test_site_watch_triggers_build(): void
+    public function test_site_serve_runs_watch_and_serves(): void
     {
         $tempOutputDir = base_path('tests/tmp_output');
 
         File::deleteDirectory($tempOutputDir);
 
         Config::set('scabbard.copy_dirs', []);
-        Config::set('scabbard.views', ['watch.html' => 'home']);
+        Config::set('scabbard.views', ['serve.html' => 'home']);
         Config::set('scabbard.output_path', $tempOutputDir);
+        Config::set('scabbard.serve_port', 5678);
 
-        Artisan::call('scabbard:watch', ['--once' => true]);
+        Artisan::call('scabbard:serve', ['--once' => true]);
 
-        $this->assertTrue(File::exists("{$tempOutputDir}/watch.html"));
+        $this->assertTrue(File::exists("{$tempOutputDir}/serve.html"));
 
         File::deleteDirectory($tempOutputDir);
     }
